requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
plotly==5.24.1
python-dotenv==1.0.1
openai==1.62.0
